Split Graphs.h/Graphs.cpp, cleanup
Technically part of one of the Graph Refactors or cleanups: Graphs.h/Graphs.cpp
contained two classes: EdgeConditionBuilder and EdgeConditionBuilderContainer.

Split those files into the two headers and only include them wehere they
are uesd.
