+++ arangod/Aql/EdgeConditionBuilderContainer.h
@@ -23,6 +23,7 @@
 
 #pragma once
 
+#include "Aql/EdgeConditionBuilder.h"
 #include "Aql/VariableGenerator.h"
 
 namespace arangodb {
@@ -34,68 +35,6 @@ class Slice;
 
 namespace aql {
 struct AstNode;
-
-// Helper class to generate AQL AstNode conditions
-// that can be handed over to Indexes in order to
-// query the data.
-// This class does NOT take responsibility
-// of the referenced AstNodes, the creater
-// has to free them.
-// In AQL the AST is responsible to free all nodes.
-class EdgeConditionBuilder {
- protected:
-  /// @brief a condition checking for _from
-  /// not used directly
-  AstNode* _fromCondition;
-
-  /// @brief a condition checking for _to
-  /// not used directly
-  AstNode* _toCondition;
-
-  /// @brief the temporary condition to ask indexes
-  /// Is always in Normalized format: One n-ary-and
-  /// branch of a DNF => No OR contained.
-  AstNode* _modCondition;
-
-  /// @brief indicator if we have attached the _from or _to condition to
-  /// _modCondition
-  bool _containsCondition;
-
-  EdgeConditionBuilder(Ast*, EdgeConditionBuilder const&);
-  explicit EdgeConditionBuilder(AstNode*,
-                                arangodb::ResourceMonitor& resourceMonitor);
-
-  // Create the _fromCondition for the first time.
-  virtual void buildFromCondition() = 0;
-
-  // Create the _toCondition for the first time.
-  virtual void buildToCondition() = 0;
-
- public:
-  virtual ~EdgeConditionBuilder() = default;
-
-  EdgeConditionBuilder(EdgeConditionBuilder const&) = delete;
-  EdgeConditionBuilder(EdgeConditionBuilder&&) = delete;
-
-  // Add a condition on the edges that is not related to
-  // the direction e.g. `label == foo`
-  void addConditionPart(AstNode const*);
-
-  // Get the complete condition for outbound edges
-  AstNode* getOutboundCondition();
-
-  // Get the complete condition for inbound edges
-  AstNode* getInboundCondition();
-
-  // Get the resource monitor
-  arangodb::ResourceMonitor& resourceMonitor();
-
- private:
-  // Internal helper to swap _from and _to parts
-  void swapSides(AstNode* condition);
-
-  arangodb::ResourceMonitor& _resourceMonitor;
-};
 
 // Wrapper around EdgeConditionBuilder that takes responsibility for all
 // AstNodes created with it. Can be used outside of an AQL query.
