+++ arangod/Graph/GraphManager.cpp
@@ -33,7 +33,6 @@
 
 #include "ApplicationFeatures/ApplicationServer.h"
 #include "Aql/AstNode.h"
-#include "Aql/Graphs.h"
 #include "Aql/Query.h"
 #include "Aql/QueryOptions.h"
 #include "Basics/ReadLocker.h"
